stt_api/requirements.txt
Django==3.1.12
djangorestframework==3.13.1
djongo==1.3.6
gunicorn==20.1.0
mongoengine==0.24.1
pymongo==3.12.3
requests==2.27.1
